Added Vector Line Preview Before Launch
Introduced a new variable (admin.draw_lines) which holds information to draw or not to draw lines. It turns True when the game is ready to launch. A clock.schedule was introduced to the launch to delay the launch. Admin.draw_lines is set back to False when they launch after the clock.